requirements.txt
Django==1.6.5
Kivy==1.7.2
Mako==0.7.3
Markdown==2.2.1
MarkupSafe==0.15
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.0.0
PyPubSub==3.1.2
PyYAML==3.11
Pygments==1.6
Skype4Py==1.0.32.0
Twisted-Core==12.3.0
Twisted-Names==12.3.0
Twisted-Web==12.3.0
adium-theme-ubuntu==0.3.3
apt-xapian-index==0.45
argparse==1.2.1
beautifulsoup4==4.3.2
chardet==2.0.1
configglue==1.0
cx-Freeze==4.3.1
debtagshw==0.1
defer==1.0.6
dirspec==4.2.0
distribute==0.6.34
dj-database-url==0.3.0
dj-static==0.0.6
django-ajax-selects==1.3.5
django-autocomplete-light==2.0.0a18
django-toolbelt==0.0.1
docutils==0.10
duplicity==0.6.21
gunicorn==19.1.1
httplib2==0.7.7
lxml==3.1.0
mercurial==2.6.3
numpy==1.7.1
oauth==1.0.1
oauth2client==1.2
oauthlib==0.6.0
oneconf==0.3.3
pexpect==2.4
piston-mini-client==0.7.5
protobuf==2.4.1
psycopg2==2.4.5
py-mysql2pgsql==0.1.5
pyOpenSSL==0.13
pycrypto==2.6
pycups==1.9.62
pycurl==7.19.0
pygame==1.9.1release
pygobject==3.8.0
pyinotify==0.9.3
pyserial==2.6
pysmbc==1.0.13
pysqlite==1.0.1
python-apt==0.8.8ubuntu6
python-debian==0.1.21-nmu2ubuntu1
python-oauth2==0.4.0
python-openid==2.2.5
python-social-auth==0.1.17
pyxdg==0.25
qrcode==4.0.4
reportlab==2.6
requests==2.1.0
requests-oauthlib==0.4.0
rhythmbox-ubuntuone==4.2.0
roman==1.4.0
sessioninstaller==0.0.0
six==1.7.3
skype-wrapper==0.6.2.0
software-center-aptd-plugins==0.0.0
static3==0.5.1
system-service==0.1.6
termcolor==1.1.0
ubuntuone-storage-protocol==4.2.0
unity-lens-photos==0.9
urlgrabber==3.9.1
wsgiref==0.1.2
wxGlade==0.6.5
wxPython==2.8.12.1
wxPython-common==2.8.12.1
xdiagnose==3.5.1
xlrd==0.9.2
xlutils==1.6.0
xlwt==0.7.5
yum-metadata-parser==1.1.4
zope.interface==4.0.5
